convert comm operand to ushort

+++ Source/AtlusScriptLibrary/FlowScriptLanguage/Decompiler/Evaluator.cs
@@ -36,7 +36,7 @@ public class Evaluator
 
     // FlowScript evaluation
     private FlowScript mScript;
-    private Dictionary<int, FunctionDeclaration> mFunctions;
+    private Dictionary<uint, FunctionDeclaration> mFunctions;
     private Dictionary<int, ProcedureDeclaration> mProcedures;
     private Stack<EvaluatedScope> mScopeStack;
     private EvaluatedScope Scope => mScopeStack.Peek();
@@ -123,7 +123,7 @@ public class Evaluator
     private void InitializeScriptEvaluationState(FlowScript flowScript)
     {
         mScript = flowScript;
-        mFunctions = new Dictionary<int, FunctionDeclaration>();
+        mFunctions = new Dictionary<uint, FunctionDeclaration>();
         mProcedures = new Dictionary<int, ProcedureDeclaration>();
         mScopeStack = new Stack<EvaluatedScope>();
     }
@@ -336,7 +336,7 @@ public class Evaluator
         LogInfo("Registering functions");
         foreach (var instruction in mScript.Procedures.SelectMany(x => x.Instructions).Where(x => x.Opcode == Opcode.COMM))
         {
-            var index = instruction.Operand.Int16Value;
+            var index = (ushort)instruction.Operand.Int16Value;
             if (mFunctions.ContainsKey(index))
                 continue;
 
@@ -467,7 +467,7 @@ public class Evaluator
                 break;
             case Opcode.COMM:
                 {
-                    short index = instruction.Operand.Int16Value;
+                    ushort index = (ushort)instruction.Operand.Int16Value;
                     foreach (var parameter in mFunctions[index].Parameters)
                     {
                         if (stack.Count != 0)
@@ -787,7 +787,7 @@ public class Evaluator
             // Call to function
             case Opcode.COMM:
                 {
-                    short index = instruction.Operand.Int16Value;
+                    ushort index = (ushort)instruction.Operand.Int16Value;
                     if (!mFunctions.TryGetValue(index, out var function))
                     {
                         LogError("Unknown function: registration of functions must have failed");
